LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-double-activation-expected.txt



FAIL CSS Values Test: ::scroll-marker double activation scrolls into view twice assert_equals: ::scroll-marker activation retains focus expected "rgb(0, 0, 255)" but got ""
